Make discriminated union code generic
