+++ wagtail/api/v3/routers/pages.py
@@ -9,14 +9,16 @@ from wagtail.api.v3.form_data import build_page_form, build_page_update_form
 from wagtail.api.v3.pagination import WagtailLimitOffsetPagination
 from wagtail.api.v3.permissions import require_any_permission
 from wagtail.api.v3.querysets import AccessTier, get_pages_queryset
-from wagtail.api.v3.schemas import BasePageSchema, build_page_schema_unions
+from wagtail.api.v3.schemas import BasePageSchema
+from wagtail.api.v3.schemas.base import build_union_schemas
 from wagtail.coreutils import resolve_model_string
 from wagtail.models import Page, get_page_models
 from wagtail.utils.forms import FormValidationError
 
 router = Router(tags=["pages"])
 
-_page_schemas = build_page_schema_unions(get_page_models())
+
+_page_schemas = build_union_schemas(get_page_models())
 PageDetailSchema = _page_schemas.detail
 PageCreateSchema = _page_schemas.create
 PageUpdateSchema = _page_schemas.update
