+++ wagtail/api/v3/schemas/__init__.py
@@ -1,14 +1,19 @@
-from .base import BaseMetaSchema, BaseSchema, ContentTypeSummarySchema
+from .base import (
+    BaseMetaSchema,
+    BaseSchema,
+    ContentTypeSummarySchema,
+    DiscriminatedUnionSchemas,
+    build_union_schemas,
+    discriminate_schema,
+)
 from .generators import create_generator, patch_generator, read_generator
 from .pages import (
     BasePageSchema,
     PageCreateBaseSchema,
     PageCreateMetaSchema,
     PageMetaSchema,
-    PageSchemaUnions,
     PageUpdateBaseSchema,
     PageUpdateMetaSchema,
-    build_page_schema_unions,
 )
 from .sites import SiteInputSchema, SiteSchema
 
@@ -16,6 +21,9 @@ __all__ = [
     "BaseMetaSchema",
     "BaseSchema",
     "ContentTypeSummarySchema",
+    "DiscriminatedUnionSchemas",
+    "build_union_schemas",
+    "discriminate_schema",
     "read_generator",
     "create_generator",
     "patch_generator",
